Add experimental OAuth support, fix double archiving of "bare" links, fix crash getting repost sub when null/blank
Config file needs to be remade, otherwise using OAuth is optional. Not
sure if it needs to refresh its token or not.
Links posted without being wrapped in brackets were being archived
twice, tweaked regex and method to avoid.
If repost sub is not set in config, bot would throw exception trying to
get the null subreddit. Should be handled now.